requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.18.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0